requirements.txt
amqp==2.6.0
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.5
certifi==2020.4.5.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
djangorestframework==3.11.0
drf-yasg==1.17.1
future==0.18.2
idna==2.9
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.10
MarkupSafe==1.1.1
packaging==20.4
Pillow==7.1.2
pkg-resources==0.0.0
pyparsing==2.4.7
pytz==2020.1
redis==3.5.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
